feat: set ID and description for the vip 387
