--- vips/vip-400/bsctestnet.ts
@@ -1,197 +1,0 @@
-import { BigNumberish } from "ethers";
-import { parseUnits } from "ethers/lib/utils";
-import { ethers } from "hardhat";
-import { NETWORK_ADDRESSES } from "src/networkAddresses";
-import { ProposalType } from "src/types";
-import { makeProposal } from "src/utils";
-
-const { VTREASURY, RESILIENT_ORACLE, REDSTONE_ORACLE, UNITROLLER, ACCESS_CONTROL_MANAGER } =
-  NETWORK_ADDRESSES.bsctestnet;
-export const PROTOCOL_SHARE_RESERVE = "0x25c7c7D6Bf710949fD7f03364E9BA19a1b3c10E3";
-const SOLVBTC = "0x6855E14A6df91b8E4D55163d068E9ef2530fd4CE";
-const SOLVBTC_VTOKEN = "0xA38110ae4451A86ab754695057d5B5a9BEAd0387";
-const FIXED_SOLVBTC_PRICE = parseUnits("60000", 18);
-const REDUCE_RESERVES_BLOCK_DELTA = "28800";
-
-const ETH = "0x98f7A83361F7Ac8765CcEBAB1425da6b341958a7";
-const USDT = "0xA11c8D9DC9b66E209Ef60F0C8D969D3CD988782c";
-const USDC = "0x16227D60f7a0e586C66B005219dfc887D13C9531";
-const BTCB = "0xA808e341e8e723DC6BA0Bb5204Bafc2330d7B8e4";
-const XVS = "0xB9e0E753630434d7863528cc73CB7AC638a7c8ff";
-const RISK_FUND_CONVERTER = "0x32Fbf7bBbd79355B86741E3181ef8c1D9bD309Bb";
-const USDT_PRIME_CONVERTER = "0xf1FA230D25fC5D6CAfe87C5A6F9e1B17Bc6F194E";
-const USDC_PRIME_CONVERTER = "0x2ecEdE6989d8646c992344fF6C97c72a3f811A13";
-const BTCB_PRIME_CONVERTER = "0x989A1993C023a45DA141928921C0dE8fD123b7d1";
-const ETH_PRIME_CONVERTER = "0xf358650A007aa12ecC8dac08CF8929Be7f72A4D9";
-const XVS_VAULT_CONVERTER = "0x258f49254C758a0E37DAb148ADDAEA851F4b02a2";
-
-export const EXPECTED_CONVERSION_INCENTIVE = 1e14;
-export const converterBaseAssets = {
-  [RISK_FUND_CONVERTER]: USDT,
-  [USDT_PRIME_CONVERTER]: USDT,
-  [USDC_PRIME_CONVERTER]: USDC,
-  [BTCB_PRIME_CONVERTER]: BTCB,
-  [ETH_PRIME_CONVERTER]: ETH,
-  [XVS_VAULT_CONVERTER]: XVS,
-};
-
-export const marketSpec = {
-  vToken: {
-    address: SOLVBTC_VTOKEN,
-    name: "Venus SolvBTC",
-    symbol: "vSolvBTC",
-    underlying: {
-      address: SOLVBTC,
-      decimals: 18,
-      symbol: "SolvBTC",
-    },
-    decimals: 8,
-    exchangeRate: parseUnits("1", 28),
-    comptroller: UNITROLLER,
-    isLegacyPool: true,
-  },
-  interestRateModel: {
-    address: "0x12b88631a2033c0e9afA36b295cdFCf49493E89e",
-    base: "0",
-    multiplier: "0.09",
-    jump: "2",
-    kink: "0.5",
-  },
-  initialSupply: {
-    amount: parseUnits("0.1572404", 18),
-    vTokenReceiver: VTREASURY,
-  },
-  riskParameters: {
-    supplyCap: parseUnits("100", 18),
-    borrowCap: parseUnits("55", 18),
-    collateralFactor: parseUnits("0.75", 18),
-    reserveFactor: parseUnits("0.2", 18),
-  },
-};
-
-const configureConverters = (fromAssets: string[], incentive: BigNumberish = EXPECTED_CONVERSION_INCENTIVE) => {
-  enum ConversionAccessibility {
-    NONE = 0,
-    ALL = 1,
-    ONLY_FOR_CONVERTERS = 2,
-    ONLY_FOR_USERS = 3,
-  }
-
-  return Object.entries(converterBaseAssets).map(([converter, baseAsset]: [string, string]) => {
-    const conversionConfigs = fromAssets.map(() => [incentive, ConversionAccessibility.ALL]);
-    return {
-      target: converter,
-      signature: "setConversionConfigs(address,address[],(uint256,uint8)[])",
-      params: [baseAsset, fromAssets, conversionConfigs],
-    };
-  });
-};
-
-export const vip400 = () => {
-  const meta = {
-    version: "v2",
-    title: "VIP-400",
-    description: ``,
-    forDescription: "I agree that Venus Protocol should proceed with this proposal",
-    againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
-    abstainDescription: "I am indifferent to whether Venus Protocol proceeds or not",
-  };
-
-  return makeProposal(
-    [
-      // Configure Oracle
-      {
-        target: REDSTONE_ORACLE,
-        signature: "setDirectPrice(address,uint256)",
-        params: [marketSpec.vToken.underlying.address, FIXED_SOLVBTC_PRICE],
-      },
-
-      {
-        target: RESILIENT_ORACLE,
-        signature: "setTokenConfig((address,address[3],bool[3]))",
-        params: [
-          [
-            marketSpec.vToken.underlying.address,
-            [REDSTONE_ORACLE, ethers.constants.AddressZero, ethers.constants.AddressZero],
-            [true, false, false],
-          ],
-        ],
-      },
-
-      // Add Market
-      {
-        target: marketSpec.vToken.comptroller,
-        signature: "_supportMarket(address)",
-        params: [marketSpec.vToken.address],
-      },
-
-      {
-        target: marketSpec.vToken.comptroller,
-        signature: "_setMarketSupplyCaps(address[],uint256[])",
-        params: [[marketSpec.vToken.address], [marketSpec.riskParameters.supplyCap]],
-      },
-
-      {
-        target: marketSpec.vToken.comptroller,
-        signature: "_setMarketBorrowCaps(address[],uint256[])",
-        params: [[marketSpec.vToken.address], [marketSpec.riskParameters.borrowCap]],
-      },
-
-      {
-        target: marketSpec.vToken.comptroller,
-        signature: "_setCollateralFactor(address,uint256)",
-        params: [marketSpec.vToken.address, marketSpec.riskParameters.collateralFactor],
-      },
-
-      {
-        target: marketSpec.vToken.address,
-        signature: "setAccessControlManager(address)",
-        params: [ACCESS_CONTROL_MANAGER],
-      },
-      {
-        target: marketSpec.vToken.address,
-        signature: "setProtocolShareReserve(address)",
-        params: [PROTOCOL_SHARE_RESERVE],
-      },
-
-      {
-        target: marketSpec.vToken.address,
-        signature: "setReduceReservesBlockDelta(uint256)",
-        params: [REDUCE_RESERVES_BLOCK_DELTA],
-      },
-      {
-        target: marketSpec.vToken.address,
-        signature: "_setReserveFactor(uint256)",
-        params: [marketSpec.riskParameters.reserveFactor],
-      },
-
-      // Mint initial supply
-      {
-        target: marketSpec.vToken.underlying.address,
-        signature: "faucet(uint256)",
-        params: [marketSpec.initialSupply.amount],
-      },
-      {
-        target: marketSpec.vToken.underlying.address,
-        signature: "approve(address,uint256)",
-        params: [marketSpec.vToken.address, marketSpec.initialSupply.amount],
-      },
-      {
-        target: marketSpec.vToken.address,
-        signature: "mintBehalf(address,uint256)",
-        params: [marketSpec.initialSupply.vTokenReceiver, marketSpec.initialSupply.amount],
-      },
-      {
-        target: marketSpec.vToken.underlying.address,
-        signature: "approve(address,uint256)",
-        params: [marketSpec.vToken.address, 0],
-      },
-
-      ...configureConverters([marketSpec.vToken.underlying.address]),
-    ],
-    meta,
-    ProposalType.REGULAR,
-  );
-};
-
-export default vip400;
